+++ packages/bundler/src/test/html-bundler_test.ts
@@ -39,7 +39,7 @@ suite('HtmlBundler', () => {
       `,
       'module.js': heredoc`
         console.log('import/export-free code');
-      `
+      `,
     });
     const bundler = new Bundler({analyzer});
     const entrypointUrl = analyzer.resolveUrl('entrypoint.html')!;
@@ -81,7 +81,7 @@ suite('HtmlBundler', () => {
     const analyzer = new Analyzer({
       urlResolver: new FsUrlResolver(root),
       urlLoader: new FsUrlLoader(root),
-      moduleResolution: 'node'
+      moduleResolution: 'node',
     });
     const bundler = new Bundler({analyzer});
     const importUsingNodeModuleResolutionUrl =
@@ -225,6 +225,24 @@ suite('HtmlBundler', () => {
       });
 
       suite('Resolve Paths', () => {
+        test('rewrite relative paths, not absolute paths', () => {
+          const html = `
+            <a href="/absolute/path">Absolute Path</a>
+            <a href="relative/path">Relative Path</a>
+          `;
+
+          const expected = `
+            <a href="/absolute/path">Absolute Path</a>
+            <a href="my-element/relative/path">Relative Path</a>
+          `;
+
+          const ast = parse(html);
+          htmlBundler['_rewriteAstBaseUrl'](ast, importDocUrl, mainDocUrl);
+
+          const actual = parse5.serialize(ast);
+          assert.deepEqual(stripSpace(actual), stripSpace(expected));
+        });
+
         test('excluding template elements', () => {
           const html = `
             <link rel="import" href="../polymer/polymer.html">